+++ net/table_net_dns_record.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net"
 	"strings"
-	"time"
 
 	"github.com/miekg/dns"
 
@@ -28,18 +27,18 @@ func tableNetDNSRecord(ctx context.Context) *plugin.Table {
 		},
 		Columns: []*plugin.Column{
 			{Name: "domain", Type: proto.ColumnType_STRING, Description: "Domain name for the record."},
-			{Name: "type", Type: proto.ColumnType_STRING, Description: "Type of the DNS record: A, CNAME, MX, etc"},
-			{Name: "dns_server", Type: proto.ColumnType_STRING, Description: "Type of the DNS record: A, CNAME, MX, etc", Transform: transform.FromQual("dns_server")},
+			{Name: "type", Type: proto.ColumnType_STRING, Description: "Type of the DNS record: A, CNAME, MX, etc."},
+			{Name: "dns_server", Type: proto.ColumnType_STRING, Description: "DNS server name and port used for queries.", Transform: transform.FromQual("dns_server")},
 			{Name: "ip", Transform: transform.FromField("IP"), Type: proto.ColumnType_IPADDR, Description: "IP address for the record, such as for A records."},
 			{Name: "target", Type: proto.ColumnType_STRING, Description: "Target of the record, such as the target address for CNAME records."},
 			{Name: "priority", Type: proto.ColumnType_INT, Description: "Priority of the record, such as for MX records."},
 			{Name: "value", Type: proto.ColumnType_STRING, Description: "Value of the record, such as the text of a TXT record."},
 			{Name: "ttl", Transform: transform.FromField("TTL"), Type: proto.ColumnType_INT, Description: "Time To Live in seconds for the record in DNS cache."},
-			{Name: "serial", Type: proto.ColumnType_INT, Description: "Serial number of the record, such as for SOA records."},
-			{Name: "min_ttl", Type: proto.ColumnType_INT, Transform: transform.FromField("MinTTL"), Description: "Specifies the hostmaster email address."},
-			{Name: "refresh", Type: proto.ColumnType_INT, Description: "Specifies the SOA refresh interval. The value configures how often a name server should check it's primary server to see if there has been any updates to the zone which it does by comparing Serial numbers."},
-			{Name: "retry", Type: proto.ColumnType_INT, Description: "Specifies SOA retry value. The value indicates how long a name server should wait to retry an attempt to get fresh zone data from the primary name server if the first attempt should fail."},
-			{Name: "expire", Type: proto.ColumnType_INT, Description: "Specifies SOA expire value. A name server will no longer consider itself Authoritative if it hasn't been able to refresh the zone data in the time limit declared in this value."},
+			{Name: "serial", Type: proto.ColumnType_INT, Description: "Specifies the SOA serial number."},
+			{Name: "minimum", Type: proto.ColumnType_INT, Description: "Specifies the SOA minimum value in seconds, which indicates how long negative answers are stored in the DNS cache."},
+			{Name: "refresh", Type: proto.ColumnType_INT, Description: "Specifies the SOA refresh interval in seconds, which configures how often a name server should check its primary server to see if there has been any updates to the zone which it does by comparing Serial numbers."},
+			{Name: "retry", Type: proto.ColumnType_INT, Description: "Specifies SOA retry value in seconds, which indicates how long a name server should wait to retry an attempt to get fresh zone data from the primary name server if the first attempt should fail."},
+			{Name: "expire", Type: proto.ColumnType_INT, Description: "Specifies SOA expire value in seconds, which indicates when the zone data is no longer authoritative."},
 		},
 	}
 }
@@ -54,23 +53,10 @@ type tableDNSRecordRow struct {
 	Priority  uint16
 	Value     string
 	Serial    uint32
-	MinTTL    uint32
+	Minimum   uint32
 	Refresh   uint32
 	Retry     uint32
 	Expire    uint32
-}
-
-func getDomainQuals(domainQualsWrapper *proto.Quals) []string {
-	var domains []string
-	domainQuals := domainQualsWrapper.Quals[0].Value
-	if qualList := domainQuals.GetListValue(); qualList != nil {
-		for _, q := range qualList.Values {
-			domains = append(domains, q.GetStringValue())
-		}
-	} else {
-		domains = append(domains, domainQuals.GetStringValue())
-	}
-	return domains
 }
 
 func getTypeQuals(typeQualsWrapper *proto.Quals) []string {
@@ -113,7 +99,7 @@ func dnsTypeToDNSLibTypeEnum(recordType string) (uint16, error) {
 	case "TXT":
 		return dns.TypeTXT, nil
 	}
-	return dns.TypeANY, fmt.Errorf("Unsupported DNS record type: %g", recordType)
+	return dns.TypeANY, fmt.Errorf("Unsupported DNS record type: %s", recordType)
 }
 
 func getRecords(domain string, dnsType string, answer dns.RR) []tableDNSRecordRow {
@@ -176,7 +162,7 @@ func getRecords(domain string, dnsType string, answer dns.RR) []tableDNSRecordRo
 			Target:  typedRecord.Ns,
 			TTL:     typedRecord.Hdr.Ttl,
 			Serial:  typedRecord.Serial,
-			MinTTL:  typedRecord.Minttl,
+			Minimum: typedRecord.Minttl,
 			Refresh: typedRecord.Refresh,
 			Retry:   typedRecord.Retry,
 			Expire:  typedRecord.Expire,
@@ -222,12 +208,11 @@ func tableDNSRecordList(ctx context.Context, d *plugin.QueryData, _ *plugin.Hydr
 	c.SingleInflight = true
 	// Use our configuration for the timeout
 	c.Timeout = GetConfigTimeout(ctx, d)
-	c.DialTimeout = 10 * time.Second
-	c.ReadTimeout = 10 * time.Second
 
 	var dnsServer string
 	if d.KeyColumnQuals["dns_server"] != nil {
 		dnsServer = d.KeyColumnQualString("dns_server")
+		// Append port if not specified
 		if !strings.HasSuffix(dnsServer, ":53") {
 			dnsServer = net.JoinHostPort(dnsServer, "53")
 		}
@@ -238,7 +223,7 @@ func tableDNSRecordList(ctx context.Context, d *plugin.QueryData, _ *plugin.Hydr
 	logger.Trace("tableDNSRecordList", "Cols", queryCols)
 	logger.Trace("tableDNSRecordList", "Domain", domain)
 	logger.Trace("tableDNSRecordList", "Types", types)
-	logger.Trace("tableDNSRecordList", "DNS Server", dnsServer)
+	logger.Trace("tableDNSRecordList", "DNS server", dnsServer)
 
 	for _, dnsType := range types {
 		dnsTypeEnumVal, err := dnsTypeToDNSLibTypeEnum(dnsType)